requirements.txt
streamlit
pandas
numpy
openpyxl==3.1.4
matplotlib.pyplot
